travail sur les mocks et intégration AngularJS

+++ src/main/java/co/simplon/PoleEmploi/patrimoine/endpoint/MonumentResource.java
@@ -27,6 +27,10 @@ public class MonumentResource {
 
 	@Inject
 	private MonumentDao monumentDao;
+	
+	public void setMonumentDao(MonumentDao monumentDao) {
+		this.monumentDao = monumentDao;
+	}
 
 	@GET
 	@Produces(MediaType.APPLICATION_JSON)
@@ -64,6 +68,11 @@ public class MonumentResource {
 	@DELETE
 	@Path("{id}")
 	public void deleteMonumentById(@PathParam("id") Long id) {
+		
+		
+		if (id == null){
+			throw new IllegalArgumentException("L'id doit être renseigné.");
+		}
 		monumentDao.deleteMonumentById(id);
 	}
 }
